code/requirements.txt
paramiko==3.5.0
frida==16.5.6
tidevice3==0.11.3
opencv-python==4.10.0.84
paddlepaddle==2.6.2
paddleocr==2.9.1
httpx[socks]==0.27.2
openai==1.54.3
pydantic==2.9.2
xmljson==0.2.1
tidevice==0.12.10
pyOpenSSL==24.3.0
pyasn1==0.6.1
Appium-Python-Client==4.4.0
tiktoken==0.8.0
torch==2.5.1
torchvision==0.20.1
pandas==2.2.3
openpyxl==3.1.5
natsort==8.4.0